[Refactor] #963 Separated equipment-info.cpp/h from player-status.cpp/h
