+++ Hengband/Hengband/Hengband.vcxproj
@@ -269,6 +269,7 @@
     <ClCompile Include="..\..\src\action\movement-execution.cpp" />
     <ClCompile Include="..\..\src\main-win\graphics-win.cpp" />
     <ClCompile Include="..\..\src\player-info\alignment.cpp" />
+    <ClCompile Include="..\..\src\player-info\equipment-info.cpp" />
     <ClCompile Include="..\..\src\player-status\player-energy.cpp" />
     <ClCompile Include="..\..\src\store\cmd-store.cpp" />
     <ClCompile Include="..\..\src\cmd-io\cmd-floor.cpp" />
@@ -939,6 +940,7 @@
     <ClInclude Include="..\..\src\main-win\graphics-win.h" />
     <ClInclude Include="..\..\src\main-win\string-win.h" />
     <ClInclude Include="..\..\src\player-info\alignment.h" />
+    <ClInclude Include="..\..\src\player-info\equipment-info.h" />
     <ClInclude Include="..\..\src\player-status\player-energy.h" />
     <ClInclude Include="..\..\src\player-status\player-hand-types.h" />
     <ClInclude Include="..\..\src\store\cmd-store.h" />
