+++ src/Makefile.am
@@ -646,6 +646,7 @@ hengband_SOURCES = \
 	player-info/base-status-info.cpp player-info/base-status-info.h \
 	player-info/body-improvement-info.cpp player-info/body-improvement-info.h \
 	player-info/class-ability-info.cpp player-info/class-ability-info.h \
+	player-info/equipment-info.cpp player-info/equipment-info.h \
 	player-info/mutation-info.cpp player-info/mutation-info.h \
 	player-info/alignment.cpp player-info/alignment.h \
 	player-info/race-ability-info.cpp player-info/race-ability-info.h \
