+++ src/action/weapon-shield.cpp
@@ -11,8 +11,8 @@
 #include "object-hook/hook-checker.h"
 #include "object-hook/hook-weapon.h"
 #include "object/object-generator.h"
+#include "player-info/equipment-info.h"
 #include "player-status/player-hand-types.h"
-#include "player/player-status.h"
 #include "system/object-type-definition.h"
 #include "system/player-type-definition.h"
 #include "view/display-messages.h"
