+++ src/mind/mind-elementalist.cpp
@@ -37,10 +37,10 @@
 #include "monster-race/race-flags7.h"
 #include "monster/monster-describer.h"
 #include "player-info/avatar.h"
+#include "player-info/equipment-info.h"
 #include "player-status/player-energy.h"
 #include "player-status/player-status-base.h"
 #include "player/player-status-table.h"
-#include "player/player-status.h"
 #include "racial/racial-util.h"
 #include "spell-kind/earthquake.h"
 #include "spell-kind/magic-item-recharger.h"
