+++ src/mind/mind-ninja.cpp
@@ -8,9 +8,9 @@
 #include "effect/effect-processor.h"
 #include "effect/spells-effect-util.h"
 #include "floor/cave.h"
-#include "floor/geometry.h"
 #include "floor/floor-object.h"
 #include "floor/floor-util.h"
+#include "floor/geometry.h"
 #include "game-option/disturbance-options.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
@@ -28,10 +28,10 @@
 #include "object/object-generator.h"
 #include "object/object-kind-hook.h"
 #include "player-attack/player-attack-util.h"
-#include "player/attack-defense-types.h"
+#include "player-info/equipment-info.h"
 #include "player-status/player-energy.h"
+#include "player/attack-defense-types.h"
 #include "player/player-status-flags.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "spell-kind/spells-detection.h"
 #include "spell-kind/spells-fetcher.h"
