+++ src/mspell/mspell-damage-calculator.cpp
@@ -1,11 +1,11 @@
 ﻿#include "mspell/mspell-damage-calculator.h"
-#include "inventory/inventory-slot-types.h"
 #include "game-option/birth-options.h"
+#include "inventory/inventory-slot-types.h"
 #include "monster-race/monster-race.h"
 #include "monster-race/race-ability-flags.h"
 #include "monster-race/race-flags2.h"
 #include "monster/monster-status.h"
-#include "player/player-status.h"
+#include "player-info/equipment-info.h"
 #include "system/floor-type-definition.h"
 #include "system/monster-race-definition.h"
 #include "system/monster-type-definition.h"
