+++ src/player-info/alignment.cpp
@@ -2,13 +2,13 @@
 #include "artifact/fixed-art-types.h"
 #include "game-option/text-display-options.h"
 #include "inventory/inventory-slot-types.h"
-#include "monster/monster-info.h"
 #include "monster-race/monster-race.h"
 #include "monster-race/race-flags3.h"
+#include "monster/monster-info.h"
 #include "monster/monster-status.h"
 #include "player-info/avatar.h"
+#include "player-info/equipment-info.h"
 #include "player/player-race.h"
-#include "player/player-status.h" // todo has_melee_weapon()が相互依存している。後で消す
 #include "system/floor-type-definition.h"
 #include "system/monster-race-definition.h"
 #include "system/monster-type-definition.h"
