+++ src/player/patron.cpp
@@ -13,10 +13,10 @@
 #include "mutation/mutation-investor-remover.h"
 #include "object-enchant/object-curse.h"
 #include "object/object-kind-hook.h"
+#include "player-info/equipment-info.h"
 #include "player/player-class.h"
 #include "player/player-damage.h"
 #include "player/player-race-types.h"
-#include "player/player-status.h"
 #include "spell-kind/spells-floor.h"
 #include "spell-kind/spells-genocide.h"
 #include "spell-kind/spells-launcher.h"
