+++ src/player/player-status-flags.cpp
@@ -13,11 +13,12 @@
 #include "object-hook/hook-checker.h"
 #include "object-hook/hook-weapon.h"
 #include "object/object-flags.h"
+#include "player-info/equipment-info.h"
 #include "player-status/player-basic-statistics.h"
-#include "player-status/player-speed.h"
-#include "player-status/player-stealth.h"
 #include "player-status/player-hand-types.h"
 #include "player-status/player-infravision.h"
+#include "player-status/player-speed.h"
+#include "player-status/player-stealth.h"
 #include "player/attack-defense-types.h"
 #include "player/digestion-processor.h"
 #include "player/mimic-info-table.h"
