+++ src/racial/racial-switcher.cpp
@@ -46,6 +46,7 @@
 #include "mind/stances-table.h"
 #include "mutation/mutation-flag-types.h"
 #include "object/item-tester-hooker.h"
+#include "player-info/equipment-info.h"
 #include "player-status/player-energy.h"
 #include "player-status/player-hand-types.h"
 #include "player/attack-defense-types.h"
