+++ src/realm/realm-hissatsu.cpp
@@ -30,9 +30,9 @@
 #include "monster/monster-update.h"
 #include "object-enchant/tr-types.h"
 #include "object/object-flags.h"
+#include "player-info/equipment-info.h"
 #include "player/player-damage.h"
 #include "player/player-move.h"
-#include "player/player-status.h"
 #include "spell-kind/earthquake.h"
 #include "spell-kind/spells-detection.h"
 #include "spell-kind/spells-launcher.h"
