+++ src/spell-realm/spells-craft.cpp
@@ -14,9 +14,9 @@
 #include "object-hook/hook-enchant.h"
 #include "object/item-use-flags.h"
 #include "object/object-flags.h"
-#include "player/attack-defense-types.h"
 #include "player-info/avatar.h"
-#include "player/player-status.h"
+#include "player-info/equipment-info.h"
+#include "player/attack-defense-types.h"
 #include "player/special-defense-types.h"
 #include "racial/racial-android.h"
 #include "spell/spells-object.h"
