+++ src/view/display-player-middle.cpp
@@ -9,6 +9,7 @@
 #include "perception/object-perception.h"
 #include "player-status/player-hand-types.h"
 #include "player/attack-defense-types.h"
+#include "player-info/equipment-info.h"
 #include "player/player-race-types.h"
 #include "player/player-skill.h"
 #include "player/player-status-flags.h"
