+++ src/view/status-first-page.cpp
@@ -17,9 +17,9 @@
 #include "object-enchant/tr-types.h"
 #include "object/object-flags.h"
 #include "perception/object-perception.h"
+#include "player-info/equipment-info.h"
 #include "player-status/player-hand-types.h"
 #include "player/player-status-flags.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "sv-definition/sv-weapon-types.h"
 #include "system/object-type-definition.h"
